requirements.txt
torch
transformers
sentence-transformers
pdfplumber
faiss-cpu
langchain
streamlit
speechrecognition
pyaudio
numpy
